Create `Identity.asRole`
This moves all session creations into the service.
