+++ src/components/workflow/permission.serializer.ts
@@ -1,7 +1,6 @@
 import { setOf } from '@seedcompany/common';
-import { DateTime } from 'luxon';
-import { type ID, Role, type Session } from '~/common';
-import { type SessionHost } from '../authentication';
+import { type ID, Role } from '~/common';
+import { type Identity } from '~/core/authentication';
 import { type Privileges, type UserResourcePrivileges } from '../authorization';
 import { Condition } from '../authorization/policy/conditions';
 import { type Workflow } from './define-workflow';
@@ -12,18 +11,11 @@ export const transitionPermissionSerializer =
   <W extends Workflow>(
     workflow: W,
     privileges: Privileges,
-    sessionHost: SessionHost,
+    identity: Identity,
   ) =>
   (transition: W['transition']): readonly SerializedTransitionPermission[] => {
     const all = [...Role].flatMap((role) => {
-      const session: Session = {
-        token: 'system',
-        issuedAt: DateTime.now(),
-        userId: 'anonymous' as ID,
-        anonymous: false,
-        roles: [`global:${role}`],
-      };
-      return sessionHost.withSession(session, () => {
+      return identity.asRole(role, () => {
         const p = privileges.for(workflow.eventResource);
         const readEvent = resolve(p, 'read', transition.key);
         const execute = resolve(p, 'create', transition.key);
